+++ packages/bitswap/test/network.spec.ts
@@ -158,7 +158,8 @@ describe('network', () => {
       id: peerId,
       multiaddrs: [
         multiaddr('/ip4/127.0.0.1/tcp/4001')
-      ]
+      ],
+      routing: 'test-routing'
     }]
 
     components.routing.findProviders.withArgs(cid).returns((async function * () {
@@ -181,7 +182,8 @@ describe('network', () => {
       id: peerId,
       multiaddrs: [
         multiaddr('/ip4/127.0.0.1/tcp/4001/p2p-circuit')
-      ]
+      ],
+      routing: 'test-routing'
     }]
 
     components.routing.findProviders.withArgs(cid).returns((async function * () {
@@ -216,7 +218,8 @@ describe('network', () => {
       id: peerId,
       multiaddrs: [
         multiaddr('/ip4/127.0.0.1/tcp/4001/p2p-circuit')
-      ]
+      ],
+      routing: 'test-routing'
     }]
 
     components.routing.findProviders.withArgs(cid).returns((async function * () {
@@ -239,7 +242,8 @@ describe('network', () => {
       id: peerId,
       multiaddrs: [
         multiaddr('/ip4/127.0.0.1/tcp/4001')
-      ]
+      ],
+      routing: 'test-routing'
     }]
 
     components.routing.findProviders.withArgs(cid).returns((async function * () {
@@ -298,7 +302,8 @@ describe('network', () => {
       id: peerId,
       multiaddrs: [
         multiaddr('/ip4/127.0.0.1/tcp/4001/p2p-circuit')
-      ]
+      ],
+      routing: 'test-routing'
     }]
 
     components.routing.findProviders.withArgs(cid).returns((async function * () {