+++ packages/block-brokers/src/trustless-gateway/utils.ts
@@ -1,13 +1,16 @@
 import { getNetConfig, isPrivate } from '@libp2p/utils'
 import { DNS, HTTP, HTTPS } from '@multiformats/multiaddr-matcher'
 import { multiaddrToUri } from '@multiformats/multiaddr-to-uri'
+import { CustomProgressEvent } from 'progress-events'
 import { Uint8ArrayList } from 'uint8arraylist'
 import { TrustlessGateway } from './trustless-gateway.js'
+import type { TrustlessGatewayGetBlockProgressEvents, TrustlessGatewayProvider } from './index.ts'
 import type { TransformRequestInit } from './trustless-gateway.js'
 import type { Routing } from '@helia/interface'
 import type { ComponentLogger, Logger, AbortOptions } from '@libp2p/interface'
 import type { Multiaddr } from '@multiformats/multiaddr'
 import type { CID } from 'multiformats/cid'
+import type { ProgressOptions } from 'progress-events'
 
 export function filterNonHTTPMultiaddrs (multiaddrs: Multiaddr[], allowInsecure: boolean, allowLocal: boolean): Multiaddr[] {
   return multiaddrs.filter(ma => {
@@ -36,7 +39,7 @@ export function filterNonHTTPMultiaddrs (multiaddrs: Multiaddr[], allowInsecure:
   })
 }
 
-export interface FindHttpGatewayProvidersOptions extends AbortOptions {
+export interface FindHttpGatewayProvidersOptions extends AbortOptions, ProgressOptions<TrustlessGatewayGetBlockProgressEvents> {
   transformRequestInit?: TransformRequestInit
 }
 
@@ -54,6 +57,15 @@ export async function * findHttpGatewayProviders (cid: CID, routing: Routing, lo
     // /ip4/x.x.x.x/tcp/31337/https
     // etc
     const uri = multiaddrToUri(httpAddresses[0])
+
+    const prov: TrustlessGatewayProvider = {
+      type: 'trustless-gateway',
+      cid,
+      url: uri.toString(),
+      routing: provider.routing
+    }
+
+    options?.onProgress?.(new CustomProgressEvent('trustless-gateway:found-provider', prov))
 
     yield new TrustlessGateway(uri, { logger, transformRequestInit: options.transformRequestInit })
   }