+++ packages/utils/src/graph-walker.ts
@@ -1,4 +1,5 @@
 import { Queue } from '@libp2p/utils'
+import filter from 'it-filter'
 import toBuffer from 'it-to-buffer'
 import { createUnsafe } from 'multiformats/block'
 import type { CodecLoader } from '@helia/interface'
@@ -22,7 +23,7 @@ export interface GraphNode <T = unknown, C extends number = number, A extends nu
 }
 
 export interface GraphWalker {
-  walk <T = any> (cid: CID, options?: AbortOptions): AsyncGenerator<GraphNode<T>>
+  walk <T = any> (cid: CID, options?: WalkOptions<T>): AsyncGenerator<GraphNode<T>>
 }
 
 export function depthFirstWalker (components: GraphWalkerComponents, init: GraphWalkerInit = {}): GraphWalker {
@@ -39,15 +40,88 @@ interface JobOptions extends AbortOptions {
   path: CID[]
 }
 
-class DepthFirstGraphWalker {
+export interface WalkOptions<T> extends AbortOptions {
+  includeChild?(child: CID, parent: BlockView<T, number, number, 0 | 1>): boolean
+}
+
+abstract class AbstractGraphWalker {
   private readonly components: GraphWalkerComponents
 
-  constructor (components: GraphWalkerComponents, init: GraphWalkerInit = {}) {
+  constructor (components: GraphWalkerComponents, init: GraphWalkerInit) {
     this.components = components
   }
 
-  async * walk <T = any> (cid: CID, options: AbortOptions): AsyncGenerator<GraphNode<T>> {
-    const queue = new Queue<GraphNode<T>, JobOptions>({
+  async * walk <T = any> (cid: CID, options?: WalkOptions<T>): AsyncGenerator<GraphNode<T>> {
+    const queue = this.getQueue()
+    const gen = filter(queue.toGenerator(options), (node) => node != null) as AsyncGenerator<GraphNode<T>>
+    let finished = false
+
+    const job = async (opts: JobOptions): Promise<GraphNode<T> | undefined> => {
+      const cid = opts.cid
+      const bytes = await toBuffer(this.components.blockstore.get(cid, opts))
+      const block = createUnsafe<T, number, number, 0 | 1>({
+        cid,
+        bytes,
+        codec: await this.components.getCodec(cid.code)
+      })
+
+      for (const [, linkedCid] of block.links()) {
+        if (options?.includeChild?.(linkedCid, block) === false) {
+          continue
+        }
+
+        queue.add(job, {
+          ...opts,
+          cid: linkedCid,
+          depth: opts.depth + 1,
+          path: [...opts.path, linkedCid]
+        })
+          // eslint-disable-next-line no-loop-func
+          .catch(err => {
+            // only throw if the generator is still yielding results, otherwise
+            // it can cause unhandled promise rejections
+            if (!finished) {
+              gen.throw(err)
+            }
+          })
+      }
+
+      return {
+        block,
+        depth: opts.depth,
+        path: opts.path
+      }
+    }
+
+    queue.add(job, {
+      ...options,
+      cid,
+      depth: 0,
+      path: [cid]
+    })
+      .catch(err => {
+        // only throw if the generator is still yielding results, otherwise it
+        // can cause unhandled promise rejections
+        if (!finished) {
+          gen.throw(err)
+        }
+      })
+
+    try {
+      yield * gen
+    } finally {
+      finished = true
+      // abort any in-progress operations
+      queue.abort()
+    }
+  }
+
+  abstract getQueue <T> (): Queue<GraphNode<T> | undefined, JobOptions>
+}
+
+class DepthFirstGraphWalker extends AbstractGraphWalker {
+  getQueue<T>(): Queue<GraphNode<T> | undefined, JobOptions> {
+    return new Queue<GraphNode<T> | undefined, JobOptions>({
       concurrency: 1,
       sort: (a, b) => {
         if (a.options.depth === b.options.depth) {
@@ -61,62 +135,12 @@ class DepthFirstGraphWalker {
         return -1
       }
     })
-
-    const gen = queue.toGenerator()
-
-    const job = async (options: JobOptions): Promise<GraphNode<T>> => {
-      const cid = options.cid
-      const bytes = await toBuffer(this.components.blockstore.get(cid, options))
-      const block = createUnsafe({
-        cid,
-        bytes,
-        codec: await this.components.getCodec(cid.code)
-      })
-
-      for (const [, linkedCid] of block.links()) {
-        queue.add(job, {
-          ...options,
-          cid: linkedCid,
-          depth: options.depth + 1,
-          path: [...options.path, linkedCid]
-        })
-          .catch(err => {
-            gen.throw(err)
-            queue.abort()
-          })
-      }
-
-      return {
-        block,
-        depth: options.depth,
-        path: options.path
-      }
-    }
-
-    queue.add(job, {
-      ...options,
-      cid,
-      depth: 0,
-      path: [cid]
-    })
-      .catch(err => {
-        gen.throw(err)
-        queue.abort()
-      })
-
-    yield * gen
   }
 }
 
-class BreadthFirstGraphWalker {
-  private readonly components: GraphWalkerComponents
-
-  constructor (components: GraphWalkerComponents, init: GraphWalkerInit = {}) {
-    this.components = components
-  }
-
-  async * walk <T = any> (cid: CID, options: AbortOptions): AsyncGenerator<GraphNode<T>> {
-    const queue = new Queue<GraphNode<T>, JobOptions>({
+class BreadthFirstGraphWalker extends AbstractGraphWalker {
+  getQueue<T>(): Queue<GraphNode<T> | undefined, JobOptions> {
+    return new Queue<GraphNode<T> | undefined, JobOptions>({
       concurrency: 1,
       sort: (a, b) => {
         if (a.options.depth === b.options.depth) {
@@ -130,49 +154,5 @@ class BreadthFirstGraphWalker {
         return 1
       }
     })
-
-    const gen = queue.toGenerator()
-
-    const job = async (options: JobOptions): Promise<GraphNode<T>> => {
-      const cid = options.cid
-      const bytes = await toBuffer(this.components.blockstore.get(cid, options))
-      const block = createUnsafe({
-        cid,
-        bytes,
-        codec: await this.components.getCodec(cid.code)
-      })
-
-      for (const [, linkedCid] of block.links()) {
-        queue.add(job, {
-          ...options,
-          cid: linkedCid,
-          depth: options.depth + 1,
-          path: [...options.path, linkedCid]
-        })
-          .catch(err => {
-            gen.throw(err)
-            queue.abort()
-          })
-      }
-
-      return {
-        block,
-        depth: options.depth,
-        path: options.path
-      }
-    }
-
-    queue.add(job, {
-      ...options,
-      cid,
-      depth: 0,
-      path: [cid]
-    })
-      .catch(err => {
-        gen.throw(err)
-        queue.abort()
-      })
-
-    yield * gen
   }
 }