+++ packages/utils/src/routing.ts
@@ -102,9 +102,13 @@ export class Routing implements RoutingInterface, Startable {
               return null
             }
 
-            return provider
+            return {
+              ...provider,
+              protocols: peer.protocols,
+              routing: peer.routing
+            }
           } catch (err) {
-            this.log.error('could not load multiaddrs for peer %p', peer.id, err)
+            this.log.error('could not load multiaddrs for peer %p - %e', peer.id, err)
             return null
           }
         }, {
@@ -112,8 +116,10 @@ export class Routing implements RoutingInterface, Startable {
           signal: options.signal
         })
           .catch(err => {
-            this.log.error('could not load multiaddrs for peer %p', peer.id, err)
+            this.log.error('could not load multiaddrs for peer %p - %e', peer.id, err)
           })
+
+        continue
       }
 
       yield peer