+++ packages/ipns/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Changelog
 
+## [9.1.3](https://github.com/ipfs/helia/compare/ipns-v9.1.2...ipns-v9.1.3) (2025-10-29)
+
+
+### Bug Fixes
+
+* use libp2p provider routing field ([#889](https://github.com/ipfs/helia/issues/889)) ([d4d97b8](https://github.com/ipfs/helia/commit/d4d97b83f76be7e3b480052467408839f808e230))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/interface bumped from ^6.0.1 to ^6.0.2
+
+## [9.1.2](https://github.com/ipfs/helia/compare/ipns-v9.1.1...ipns-v9.1.2) (2025-10-27)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/interface bumped from ^6.0.0 to ^6.0.1
+
+## [9.1.1](https://github.com/ipfs/helia/compare/ipns-v9.1.0...ipns-v9.1.1) (2025-10-22)
+
+
+### Bug Fixes
+
+* update log formatting to print errors correctly ([#884](https://github.com/ipfs/helia/issues/884)) ([f35ecd1](https://github.com/ipfs/helia/commit/f35ecd1c8ad3c712d3882d0f0f2abaf0f0296ec1))
+
 ## [9.1.0](https://github.com/ipfs/helia/compare/ipns-v9.0.0...ipns-v9.1.0) (2025-10-14)