+++ packages/mfs/CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## [6.0.3](https://github.com/ipfs/helia/compare/mfs-v6.0.2...mfs-v6.0.3) (2025-10-29)
+
+
+### Bug Fixes
+
+* use libp2p provider routing field ([#889](https://github.com/ipfs/helia/issues/889)) ([d4d97b8](https://github.com/ipfs/helia/commit/d4d97b83f76be7e3b480052467408839f808e230))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/unixfs bumped from ^6.0.2 to ^6.0.3
+
+## [6.0.2](https://github.com/ipfs/helia/compare/mfs-v6.0.1...mfs-v6.0.2) (2025-10-27)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/unixfs bumped from ^6.0.1 to ^6.0.2
+
 ## [6.0.1](https://github.com/ipfs/helia/compare/mfs-v6.0.0...mfs-v6.0.1) (2025-10-09)