+++ packages/routers/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [4.0.3](https://github.com/ipfs/helia/compare/routers-v4.0.2...routers-v4.0.3) (2025-10-29)
+
+
+### Bug Fixes
+
+* use libp2p provider routing field ([#889](https://github.com/ipfs/helia/issues/889)) ([d4d97b8](https://github.com/ipfs/helia/commit/d4d97b83f76be7e3b480052467408839f808e230))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/interface bumped from ^6.0.0 to ^6.0.2
+
+## [4.0.2](https://github.com/ipfs/helia/compare/routers-v4.0.1...routers-v4.0.2) (2025-10-27)
+
+
+### Bug Fixes
+
+* add provider events to bitswap and trustless gateways ([#888](https://github.com/ipfs/helia/issues/888)) ([95d95da](https://github.com/ipfs/helia/commit/95d95dad7ff2a1e462b5a8a4f57ac40c4503f4ef))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/interface bumped from ^6.0.0 to ^6.0.1
+
 ## [4.0.1](https://github.com/ipfs/helia/compare/routers-v4.0.0...routers-v4.0.1) (2025-10-14)